moving run to device class
