+++ octopus_code/core/octopus/device.py
@@ -1,8 +1,13 @@
 from Queue import Queue
+import time
 
 # TODO: is there anything outside of the constructor that will actually be shared?
 class Device(object):
+    ''' Generic interface for a sound reactive thing, such as a GentlemanOctopus
+    or a school of fish
+    '''
     def __init__(self, control_queue=None, audio_stream_queue=None):
+        ''' inputs should be queues '''
         # Initialise queues
         if not control_queue:
             control_queue = Queue(1)
@@ -16,8 +21,40 @@ class Device(object):
         self.audio_stream_queue = audio_stream_queue
 
 
+    # TODO: Make this non-blocking
+    def run(self, timeout=0):
+        ''' Run a sound reactive thing for a given aount of time (blocking)'''
+        run_start = time.time()
+
+        # 0 Means forever-and-ever-and-ever
+        if not timeout:
+            timeout = float("inf")
+
+        # Generate the patterns
+        while time.time() - run_start < timeout:
+            loop_start = time.time()
+
+            try:
+                self.update()
+
+            except Exception as e:
+                print e
+
+            loop_end = time.time() - loop_start
+            loop_time = self.period - loop_end
+
+            # Sleep to give time for other processes
+            time.sleep(max(0, self.period - loop_end))
+
+    def update(self):
+        pass
+
+
 if __name__ == '__main__':
     control_queue = Queue(1)
     audio_stream_queue = Queue(1)
 
     device = Device(control_queue, audio_stream_queue)
+
+    print "Running device...."
+    device.run(2)
